% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtBrewer.R
\name{art.brewer}
\alias{art.brewer}
\title{Art Palette Generator}
\usage{
art.brewer(
  name,
  n,
  type = c("discrete", "continuous"),
  direction = c("f", "r"),
  override.order = FALSE
)
}
\arguments{
\item{name}{Name of Palette.}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}

\item{direction}{Sets order of colors. Default palette is "f" for "forward". If direction is "r" for "reverse", palette color order is reversed.}

\item{override.order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}
}
\value{
A vector of colors.
}
\description{
Palettes inspired by art pieces and a certain vision of colors.
Based on the work of Blake R Mills \href{https://github.com/BlakeRMills/MetBrewer}{on Github}.
Wes Anderson palettes are inspired by the colorful vision of Wes Anderson in his movies and adapted from this \href{https://wesandersonpalettes.tumblr.com}{post}.
}
\examples{
art.brewer("Unicorn")

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=art.brewer("Vallotton", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=art.brewer("Vibrant", 3))

}
\keyword{colors}
