% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtBrewer.R
\name{create.vignette}
\alias{create.vignette}
\title{Create palette vignettes}
\usage{
create.vignette()
}
\description{
Function for creating palettes vignettes.
}
\examples{
# Create vignette
create.vignette()
}
