% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtBrewer.R
\name{display_all}
\alias{display_all}
\title{View all Palettes available}
\usage{
display_all(
  n,
  sequential = FALSE,
  category = "all",
  direction = "f",
  override.order = FALSE
)
}
\arguments{
\item{n}{Number of requested colors. If n is left blank, default palette is returned.}

\item{sequential}{Should palettes displayed all at once, or one at a time. Default is all at once (FALSE).}

\item{category}{Allow to subset the palettes to display by categories. Some palettes are not categorized. Categories are "all", "art" or "wesanderson". Default is set to "all".}

\item{direction}{Sets order of colors. Default palette is "f" for "forward". If direction is "r" for "reverse", palette color order is reversed.}

\item{override.order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}
}
\description{
Function for viewing all palettes available in ArtBrewer.
}
\examples{
# All Palettes
display_all(sequential = FALSE, category = "all")

# By category
display_all(sequential = FALSE, category = "wesanderson")

# 5 Colors of all Palettes
display_all(5, sequential = FALSE, category = "art")
}
