% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedBrewer.R
\name{med.brewer}
\alias{med.brewer}
\title{Med Palette Generator}
\usage{
med.brewer(
  name,
  n,
  type = c("discrete", "continuous"),
  direction = c("f", "r"),
  override.order = FALSE
)
}
\arguments{
\item{name}{Name of Palette.}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Colorful palettes inspired by art pieces.
Based on the work of Blake R Mills \href{https://github.com/BlakeRMills/MetBrewer}{on Github}.
Wes Anderson palettes are inspired by the colorful vision of Wes Anderson in his movies and adapted from this \href{https://wesandersonpalettes.tumblr.com}{post}.
}
\examples{
med.brewer("MED")

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=med.brewer("Vallotton", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=med.brewer("Albers2", 3))

}
\keyword{colors}
