% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedBrewer.R
\name{scale_color_med_c}
\alias{scale_color_med_c}
\title{MedBrewer palettes for plotting with ggplot2}
\usage{
scale_color_med_c(name, direction = 1, ...)
}
\arguments{
\item{name}{Name of Palette.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\description{
Function for using \code{MedBrewer} colors schemes in \code{ggplot2}. Use \code{\link{scale_color_med_d}} and \code{\link{scale_fill_med_d}}
for discrete scales and \code{\link{scale_color_med_c}} and \code{\link{scale_fill_med_c}} for continuous scales.
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Sepal.Length)) +
geom_point() +
scale_color_med_c("Seurat", direction=-1)
}
