% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedBrewer.R
\name{med.brewer}
\alias{med.brewer}
\title{MED Palette Generator}
\usage{
med.brewer(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of Palette.}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
Colorful palettes inspired by art.
Based on the work of Blake R Mills \href{https://github.com/BlakeRMills/MetBrewer}{on Github}.
}
\examples{
med.brewer("MED")

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=med.brewer("Valloton", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=med.brewer("Albers2", 3))

ggplot(data=iris, aes(x=Species, y=Sepal.Width, color=Sepal.Width)) +
geom_point(size=3) +
scale_color_gradientn(colors=med.brewer("Merz"))
}
\keyword{colors}
